requirements.txt
torch==2.8.0
pillow==11.3.0
transformers==4.57.2
torchvision==0.23.0
opencv-python==4.12.0.88
accelerate==1.12.0
certifi==2025.11.12
einops==0.8.1
charset-normalizer==3.4.4
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
psutil==7.1.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
timm==1.0.22
tokenizers==0.22.1
transformers==4.57.3
urllib3==2.5.0
qwen_vl_utils==0.0.14
bitsandbytes==0.48.2
numpy==2.2.6